requirements.txt
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-openai
faiss-cpu
pypdf
langgraph
flask
flask-cors
